pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "station-master-collector"
version = "0.1.0"
description = "Endpoint collector for Station Master observability platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    { name = "Station Master Team" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Telecommunications Industry",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking :: Monitoring",
]
keywords = ["monitoring", "observability", "snmp", "syslog", "network"]
dependencies = [
    "httpx>=0.27.0",
    "pysnmp>=6.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "apscheduler>=3.10.0",
    "aiofiles>=23.2.0",
    "aiosqlite>=0.19.0",
    "cryptography>=41.0.0",
    "tenacity>=8.2.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "lz4>=4.3.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
]

[project.scripts]
station-collector = "collector.main:app"

[project.urls]
Homepage = "https://github.com/csmblade/station-master-collector"
Documentation = "https://github.com/csmblade/station-master-collector#readme"
Repository = "https://github.com/csmblade/station-master-collector"
Issues = "https://github.com/csmblade/station-master-collector/issues"

[tool.hatch.build.targets.wheel]
packages = ["collector"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
